requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==4.2.16
djangorestframework==3.15.2
pydantic==2.10.2
pydantic-core==2.27.1
sqlparse==0.5.2
typing-extensions==4.12.2
